% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_hrv.R
\name{read_multi_hrv}
\alias{read_multi_hrv}
\title{Read Multiple HRV report for Shiny}
\usage{
read_multi_hrv(file_name, file_path)
}
\arguments{
\item{file_name}{Character vector of file name}

\item{file_path}{Character vector of file path}
}
\value{
A data.frame
}
\description{
Read Multiple HRV report for Shiny
}
