% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helper.R
\name{write_custom.xlsx}
\alias{write_custom.xlsx}
\title{Custom Write Excel}
\usage{
write_custom.xlsx(list_df, file_name)
}
\arguments{
\item{list_df}{list of data.frame to write}

\item{file_name}{A character indicate file name}
}
\description{
Custom Write Excel
}
